refactor: moved github output variable handling with one level deeper
