--- action.yml
@@ -1,64 +1,0 @@
-name: k6registry
-description: k6 extension registry generator
-author: Grafana Labs
-
-branding:
-  icon: settings
-  color: purple
-
-inputs:
-  in:
-    description: input file name
-    required: true
-
-  out:
-    description: output file name
-    required: false
-
-  api:
-    description: api directory name
-    required: false
-
-  quiet:
-    description: no output, only validation
-    required: false
-
-  verbose:
-    description: verbose logging
-    required: false
-
-  loose:
-    description: skip JSON schema validation
-    required: false
-
-  lint:
-    description: enable built-in linter
-    required: false
-
-  compact:
-    description: compact instead of pretty-printed output
-    required: false
-
-  catalog:
-    description: generate catalog to the specified file
-    required: false
-
-  origin:
-    description: external registry URL for default values
-    required: false
-
-  ref:
-    description: reference output URL for change detection
-    required: false
-
-  test:
-    description: api path(s) to test after generation
-    required: false
-
-outputs:
-  changed:
-    description: "true if the output has changed compared to ref"
-
-runs:
-  using: docker
-  image: docker://ghcr.io/grafana/k6registry:v0
