+++ go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/cli/go-gh/v2 v2.11.0
 	github.com/go-git/go-git/v5 v5.12.0
 	github.com/google/go-github/v62 v62.0.0
-	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510
 	github.com/grafana/clireadme v0.1.0
 	github.com/grafana/k6lint v0.3.1
 	github.com/narqo/go-badge v0.0.0-20230821190521-c9a75c019a59
